requirements.txt
# asgiref==3.11.1
# Django==6.0.3
# sqlparse==0.5.5
# tzdata==2025.3


Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
asgiref>=3.7.2
sqlparse>=0.4.4
tzdata>=2023.3
requests>=2.31.0
python-decouple>=3.8
gunicorn>=21.2.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.9

pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0
